Timer, Game to Controller to Scene linkage
Added a time class in the model section. It has two timers. One is
periodic that fires every few seconds and one is the overarching timer.
The game model starts a timer now and periodically tells the controller
to spawn an archer in the view and move to the center.